// src/app/models/userDetail.module.ts
export class UserDetail {
  Id: number;
  fullname: string;
  firstname: string;
  middlename: string;
  lastname: string;
  gender: string;
  motherTongue: string;
  Clause: string;
  dob: Date;
  birthTime: string;
  age: number;
  height: number;
  color: string;
  qualificaton: string;
  occupation: string;
  work: string;
  earnings: string;
  physicalFitness: string;
  maritalStatus: string;
  religion: string;
  familyGod: string;
  fathername: string;
  mothername: string;
  fatherOccupation: string;
  motherOccupation: string;
  sibblingsMale: number;
  sibblingsFemale: number;
  marriedSibblingsMale: number;
  marriedSibblingsFemale: number;
  origin: string;
  native: string;
  assets: string;
  dowry: string;
  star: string;
  rasi: string;
  lagnam: string;
  directionBias: string;
  year: string;
  month: string;
  day: string;
  ra1: string;
  ra2: string;
  ra3: string;
  ra4: string;
  ra5: string;
  ra6: string;
  ra7: string;
  ra8: string;
  ra9: string;
  ra10: string;
  ra11: string;
  ra12: string;
  f1: string;
  f2: string;
  f3: string;
  f4: string;
  f5: string;
  f6: string;
  f7: string;
  f8: string;
  f9: string;
  f10: string;
  f11: string;
  f12: string;
  address: string;
  district: string;
  phone1: number;
  phone2: number;
  mailid: string;
  password: string;
  expectations: string;
  otherMessage: string;
  userImage: File;
}